+++ main/sentry/index.ts
@@ -1,0 +1,44 @@
+import * as Sentry from '@sentry/electron'
+import type { Event } from '@sentry/types'
+import store from '../store'
+
+function getCrashReportFields () {
+  const fields = ['networks', 'networksMeta', 'tokens']
+
+  return fields.reduce((extra, field) => ({ ...extra, [field]: JSON.stringify(store('main', field) || {}) }), {})
+}
+
+function getSentryExceptions (event: Event) {
+  const exceptions = event?.exception?.values || []
+  const safeExceptions = exceptions.map((exception) => {
+    const frames = exception?.stacktrace?.frames || []
+    const safeExceptionModule = ({ module = '' }) => {
+      const matches = /(.+)[\\|\/]frame[\\|\/]resources[\\|\/]app.asar[\\|\/](.+)/.exec(module)
+      if (matches && matches[2]) {
+        return `{asar}/${matches[2].replaceAll('\\', '/')}`
+      }
+      return module
+    }
+    const safeFrames = frames.map((frame) => ({ ...frame, module: safeExceptionModule(frame) }))
+    return { stacktrace: { frames: safeFrames } }
+  })
+  
+  return safeExceptions
+}
+
+export function init () {
+  Sentry.init({
+    // only use IPC from renderer process, not HTTP
+    ipcMode: Sentry.IPCMode.Classic,
+    dsn: 'https://7b09a85b26924609bef5882387e2c4dc@o1204372.ingest.sentry.io/6331069',
+    beforeSend: (evt: Event) => {
+      return {
+        ...evt,
+        exception: { values: getSentryExceptions(evt) },
+        user: { ...evt.user, ip_address: undefined }, // remove IP address
+        tags: { ...evt.tags, 'frame.instance_id': store('main.instanceId') },
+        extra: getCrashReportFields()
+      }
+    }
+  })
+}
